requirement.txt
Django==1.10.3
django-ckeditor==5.1.1
django-feedparser==0.2.1
django-grappelli==2.9.1
django-jet==1.0.2
django-tinymce==2.4.0
django-wysiwyg==0.8.0
docopt==0.4.0
feedparser==5.2.1
google-api-python-client==1.4.1
httplib2==0.9.2
mailchimp==2.0.9
oauth2client==4.0.0
Pillow==3.4.2
pyasn1==0.1.9
pyasn1-modules==0.0.8
requests==2.12.1
rsa==3.4.2
six==1.10.0
uritemplate==3.0.0
